pyproject.toml
[project]
name = "transbench"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-openai>=1.0.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.3",
]
